Put back the decimal handling changes into Sum like how we started with
