+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/expressions/aggregate/Sum.scala
@@ -62,37 +62,90 @@ case class Sum(child: Expression) extends DeclarativeAggregate with ImplicitCast
 
   private lazy val sum = AttributeReference("sum", sumDataType)()
 
+  private lazy val overflow = AttributeReference("overflow", BooleanType, false)()
+
   private lazy val zero = Literal.default(sumDataType)
 
-  override lazy val aggBufferAttributes = sum :: Nil
+  override lazy val aggBufferAttributes = sum :: overflow :: Nil
 
   override lazy val initialValues: Seq[Expression] = Seq(
-    /* sum = */ Literal.create(null, sumDataType)
+    /* sum = */ Literal.create(null, sumDataType),
+    /* overflow = */ Literal.create(false, BooleanType)
   )
 
+  /**
+   * For decimal types, update will do the following:
+   * We have a overflow flag and when it is true, it indicates overflow has happened
+   * 1. Start initial state with overflow = false, sum = null
+   * 2. Set sum to null if the value overflows else sum contains the intermediate sum
+   * 3. If overflow flag is true, keep sum as null
+   * 4. If overflow happened, then set overflow flag to true
+   */
   override lazy val updateExpressions: Seq[Expression] = {
     if (child.nullable) {
-      Seq(
-        /* sum = */
-        coalesce(coalesce(sum, zero) + child.cast(sumDataType), sum)
-      )
+      resultType match {
+        case d: DecimalType =>
+          Seq(
+            If(overflow, sum, coalesce(
+              CheckOverflow(coalesce(sum, zero) + child.cast(sumDataType), d, true), sum)),
+            overflow ||
+              coalesce(HasOverflow(coalesce(sum, zero) + child.cast(sumDataType), d), false)
+          )
+        case _ => Seq(coalesce(coalesce(sum, zero) + child.cast(sumDataType), sum), false)
+      }
     } else {
-      Seq(
-        /* sum = */
-        coalesce(sum, zero) + child.cast(sumDataType)
-      )
+      resultType match {
+        case d: DecimalType =>
+          Seq(
+            If(overflow, sum,
+              CheckOverflow(coalesce(sum, zero) + child.cast(sumDataType), d, true)),
+            overflow ||
+              coalesce(HasOverflow(coalesce(sum, zero) + child.cast(sumDataType), d), false)
+          )
+        case _ => Seq(coalesce(sum, zero) + child.cast(sumDataType), false)
+      }
     }
   }
 
+  /**
+   *
+   * Decimal handling:
+   * If any of the left or right portion of the agg buffers has the overflow flag to true,
+   * then sum is set to null else sum is added for both sum.left and sum.right
+   * and if the value overflows it is set to null.
+   * If we have already seen overflow , then set overflow to true, else check if the addition
+   * overflowed and update the overflow buffer.
+   */
   override lazy val mergeExpressions: Seq[Expression] = {
-    Seq(
-      /* sum = */
-      coalesce(coalesce(sum.left, zero) + sum.right, sum.left)
-    )
+    resultType match {
+      case d: DecimalType =>
+        Seq(
+          If(coalesce(overflow.left, false) || coalesce(overflow.right, false),
+            Literal.create(null, d),
+            coalesce(CheckOverflow(coalesce(sum.left, zero) + sum.right, d, true), sum.left)),
+          If(coalesce(overflow.left, false) || coalesce(overflow.right, false),
+            true, HasOverflow(coalesce(sum.left, zero) + sum.right, d))
+        )
+      case _ =>
+        Seq(
+          coalesce(coalesce(sum.left, zero) + sum.right, sum.left),
+          false
+        )
+    }
   }
 
+  /**
+   * Decimal handling:
+   * If overflow buffer is true, and if ansiEnabled is true then throw exception, else return null
+   * If overflow did not happen, then return the sum value
+   */
   override lazy val evaluateExpression: Expression = resultType match {
-    case d: DecimalType => CheckOverflow(sum, d, !SQLConf.get.ansiEnabled)
+    case d: DecimalType =>
+      If(EqualTo(overflow, true),
+        If(!SQLConf.get.ansiEnabled,
+          Literal.create(null, sumDataType),
+          OverflowException(resultType, "Arithmetic Operation overflow")),
+        sum)
     case _ => sum
   }
