+++ sql/core/src/test/resources/sql-tests/results/explain.sql.out
@@ -918,15 +918,15 @@ Input [2]: [key#x, val#x]
 Input [2]: [key#x, val#x]
 Keys: []
 Functions [3]: [partial_count(val#x), partial_sum(cast(key#x as bigint)), partial_count(key#x) FILTER (WHERE (val#x > 1))]
-Aggregate Attributes [3]: [count#xL, sum#xL, count#xL]
-Results [3]: [count#xL, sum#xL, count#xL]
+Aggregate Attributes [4]: [count#xL, sum#xL, overflow#x, count#xL]
+Results [4]: [count#xL, sum#xL, overflow#x, count#xL]
 
 (4) Exchange
-Input [3]: [count#xL, sum#xL, count#xL]
+Input [4]: [count#xL, sum#xL, overflow#x, count#xL]
 Arguments: SinglePartition, true, [id=#x]
 
 (5) HashAggregate [codegen id : 2]
-Input [3]: [count#xL, sum#xL, count#xL]
+Input [4]: [count#xL, sum#xL, overflow#x, count#xL]
 Keys: []
 Functions [3]: [count(val#x), sum(cast(key#x as bigint)), count(key#x)]
 Aggregate Attributes [3]: [count(val#x)#xL, sum(cast(key#x as bigint))#xL, count(key#x)#xL]
